pass build for mac static libray

+++ src/internal_def.h
@@ -2,7 +2,7 @@
  * @Author: gimphammer@gmail.com
  * @Date: 2026-05-01 12:21:50
  * @LastEditors: gimphammer@gmail.com
- * @LastEditTime: 2026-05-09 17:58:01
+ * @LastEditTime: 2026-05-10 17:18:52
  * @Copyright: Copyright (c) 2026 by gimphammer@gmail.com, All rights reserved.
  * @Description: [None]
  */
@@ -12,6 +12,7 @@
 #define INTERNAL_DEF_H
 
 #include <cstdint>
+#include <vector>
 
 namespace gz_rs_fec {
 
@@ -20,8 +21,8 @@ using Matrix2DUInt8 = std::vector<Matrix1DUint8>;
 
 // x^8+x^4+x^3+x^2+1
 extern const uint16_t kPrimitivePolynomial;
-extern const uint16_t kGF256MulTable[256][256]; 
-extern const uint16_t kMIETable[256]; //mie = multiplicative inverse element
+extern const uint8_t kGF256MulTable[256][256]; 
+extern const uint8_t kMIETable[256]; //mie = multiplicative inverse element
 
 
 // uint16_t GaloisFieldMultiply(uint16_t a, uint16_t b);
